Move search message types to /models
