+++ src/models/search/index.ts
@@ -9,10 +9,13 @@
 export * from './is-search2';
 export * from './is-valid-search';
 export * from './raw-search';
+export * from './raw-search-message-received';
+export * from './raw-search-message-sent';
 export * from './raw-search2';
 export * from './search';
 export * from './search-entries';
 export * from './search-filter';
+export * from './search-message-commands';
 export * from './search-stats';
 export * from './search-subscription';
 export * from './search2';
